dataset/scripts/embeddings/ESM/requirements.txt
torch
fair-esm
click
numpy
pandas
pyarrow